// cypress/support/commands/article.ts
import { Article } from '../../../src/entities/Article/model/types/article';

const defaultArticle = {
    id: '9999',
    title: 'TESTING ARTICLE',
    subtitle: 'БиологиЯ',
    img: 'https://klike.net/uploads/posts/2022-08/1661961079_j-58.jpg',
    views: 1022,
    createdAt: '26.02.2022',
    userId: '1',
    type: ['SCIENCE'],
    blocks: [
        {
            id: '1',
            type: 'TEXT',
            title: 'Заголовок этого блока',
            paragraphs: [
                'Программа, которую по традиции называют «Hello, world!», очень проста. Она выводит куда-либо фразу',
                '«Hello, world!», или другую подобную, средствами некоего языка.',
            ],
        },
    ],
};

export const createArticle = (article?: Article) => {
    return cy
        .request({
            method: 'POST',
            url: 'http://localhost:8000/articles',
            headers: { Authorization: '1' },
            body: article ?? defaultArticle,
        })
        .then((resp) => resp.body);
};

export const removeArticle = (articleId: string) => {
    cy.request({
        method: 'DELETE',
        url: `http://localhost:8000/articles/${articleId}`,
        headers: { Authorization: '1' },
    });
};

declare global {
    namespace Cypress {
        interface Chainable {
            createArticle(article?: Article): Chainable<Article>;
            removeArticle(articleId: string): Chainable<void>;
        }
    }
}
